Providers will now specify which Cholesky factorization to use for symmetric CSC matrices (SuiteSparse or managed alternative). Tests for both implementations have been integrated.
